Rename ConfigurationEntriesCacheTest to CommonSettingsCacheTest and update references to CommonSettingsRetriever for consistency
